% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{perform_test}
\alias{perform_test}
\title{Perform Statistical Tests}
\usage{
perform_test(
  data,
  var,
  group,
  var_type,
  is_normal = NULL,
  auto_normal = TRUE,
  adjust_method = "none"
)
}
\arguments{
\item{data}{A data frame containing the variables}

\item{var}{Variable name to test}

\item{group}{Grouping variable name}

\item{var_type}{Type of variable ("continuous" or "categorical")}

\item{is_normal}{Logical, whether the continuous variable is normally distributed}

\item{auto_normal}{Logical, whether to automatically determine normality}

\item{adjust_method}{Method for p-value adjustment in multiple comparisons}
}
\value{
A list containing test results including test type, statistic, and p-value
}
\description{
Perform Statistical Tests
}
