% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/table1_sci.R
\name{print.table1sci}
\alias{print.table1sci}
\title{Print Method for Table 1}
\usage{
\method{print}{table1sci}(x, ...)
}
\arguments{
\item{x}{Object of class table1sci}

\item{...}{Additional arguments passed to print}
}
\description{
Print Method for Table 1
}
