% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statistical_tests.R
\name{print.table1sci_test}
\alias{print.table1sci_test}
\title{Print Method for Statistical Test Results}
\usage{
\method{print}{table1sci_test}(x, ...)
}
\arguments{
\item{x}{Object of class table1sci_test}

\item{...}{Additional arguments passed to print}
}
\description{
Print Method for Statistical Test Results
}
