% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/detect_var_type.R
\name{print.table1sci_vartype}
\alias{print.table1sci_vartype}
\title{Print Method for Variable Type Detection Results}
\usage{
\method{print}{table1sci_vartype}(x, ...)
}
\arguments{
\item{x}{Object of class table1sci_vartype}

\item{...}{Additional arguments passed to print}
}
\description{
Print Method for Variable Type Detection Results
}
